requirements.txt
crossword
ipuz
puzpy
pyyaml
requests
xmlschema
